# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Health"]
      }
    },
    "/me": {
      "get": {
        "operationId": "MeController_findMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            }
          }
        },
        "tags": ["Me"]
      }
    },
    "/regions": {
      "get": {
        "operationId": "RegionsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegionDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["Regions"]
      }
    },
    "/regions/{idRegion}": {
      "get": {
        "operationId": "RegionsController_findOne",
        "parameters": [
          {
            "name": "idRegion",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionDto"
                }
              }
            }
          }
        },
        "tags": ["Regions"]
      },
      "patch": {
        "operationId": "RegionsController_update",
        "parameters": [
          {
            "name": "idRegion",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRegionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionDto"
                }
              }
            }
          }
        },
        "tags": ["Regions"]
      }
    },
    "/regions/{idRegion}/sectors": {
      "get": {
        "operationId": "SectorsController_findByRegion",
        "parameters": [
          {
            "name": "idRegion",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SectorDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["Sectors"]
      }
    },
    "/sectors/{idSector}": {
      "get": {
        "operationId": "SectorController_findOne",
        "parameters": [
          {
            "name": "idSector",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectorDto"
                }
              }
            }
          }
        },
        "tags": ["Sector"]
      },
      "patch": {
        "operationId": "SectorController_update",
        "parameters": [
          {
            "name": "idSector",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSectorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectorDto"
                }
              }
            }
          }
        },
        "tags": ["Sector"]
      }
    },
    "/sectors/{idSector}/routes": {
      "get": {
        "operationId": "RoutesController_findBySector",
        "parameters": [
          {
            "name": "idSector",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RouteDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["Routes"]
      }
    },
    "/routes/{idRoute}": {
      "get": {
        "operationId": "RouteController_findOne",
        "parameters": [
          {
            "name": "idRoute",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteDto"
                }
              }
            }
          }
        },
        "tags": ["Route"]
      },
      "patch": {
        "operationId": "RouteController_update",
        "parameters": [
          {
            "name": "idRoute",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRouteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteDto"
                }
              }
            }
          }
        },
        "tags": ["Route"]
      }
    },
    "/ticks": {
      "get": {
        "operationId": "TicksController_findMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TickDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["Ticks"]
      },
      "post": {
        "operationId": "TicksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTickDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TickDto"
                }
              }
            }
          }
        },
        "tags": ["Ticks"]
      }
    }
  },
  "info": {
    "title": "crag-atlas API",
    "description": "",
    "version": "0.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "MeDto": {
        "type": "object",
        "properties": {
          "idUser": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean",
            "description": "Mirrors the user_roles row the RLS checks"
          }
        },
        "required": ["idUser", "isAdmin"]
      },
      "RegionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "rockType": {
            "type": "string"
          },
          "sectorCount": {
            "type": "number"
          },
          "routeCount": {
            "type": "number"
          },
          "gradeRange": {
            "type": "string",
            "nullable": true,
            "description": "Derived from the routes of the region; null when it has none"
          }
        },
        "required": [
          "id",
          "name",
          "province",
          "rockType",
          "sectorCount",
          "routeCount"
        ]
      },
      "UpdateRegionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "rockType": {
            "type": "string"
          }
        },
        "required": ["name", "province", "rockType"]
      },
      "SectorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "idRegion": {
            "type": "string"
          },
          "regionName": {
            "type": "string",
            "description": "Label for breadcrumbs; ids carry no meaning"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "approachMinutes": {
            "type": "number",
            "nullable": true
          },
          "routeCount": {
            "type": "number"
          },
          "gradeRange": {
            "type": "string",
            "nullable": true,
            "description": "Derived from the routes of the sector; null when it has none"
          }
        },
        "required": [
          "id",
          "idRegion",
          "regionName",
          "name",
          "description",
          "routeCount"
        ]
      },
      "UpdateSectorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "approachMinutes": {
            "type": "number",
            "nullable": true
          }
        },
        "required": ["name"]
      },
      "RouteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "idSector": {
            "type": "string"
          },
          "sectorName": {
            "type": "string",
            "description": "Label for breadcrumbs; ids carry no meaning"
          },
          "idRegion": {
            "type": "string"
          },
          "regionName": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "grade": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["sport", "trad", "boulder"]
          },
          "length": {
            "type": "number",
            "nullable": true
          },
          "boltsCount": {
            "type": "number",
            "nullable": true
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "idSector",
          "sectorName",
          "idRegion",
          "regionName",
          "name",
          "grade",
          "type",
          "description"
        ]
      },
      "UpdateRouteDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "grade": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["sport", "trad", "boulder"]
          },
          "length": {
            "type": "number",
            "nullable": true
          },
          "boltsCount": {
            "type": "number",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["name", "grade", "type"]
      },
      "TickDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "idUser": {
            "type": "string"
          },
          "idRoute": {
            "type": "string"
          },
          "routeName": {
            "type": "string",
            "nullable": true,
            "description": "Resolved from the route catalog; null if the route is gone"
          },
          "routeGrade": {
            "type": "string",
            "nullable": true
          },
          "sectorName": {
            "type": "string",
            "nullable": true
          },
          "ascentStyle": {
            "type": "string",
            "enum": [
              "onsight",
              "flash",
              "retro_flash",
              "redpoint",
              "toprope",
              "attempt"
            ]
          },
          "climbedAt": {
            "type": "string",
            "description": "ISO date, e.g. 2026-09-19"
          },
          "attempts": {
            "type": "number",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "idUser",
          "idRoute",
          "ascentStyle",
          "climbedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateTickDto": {
        "type": "object",
        "properties": {
          "idRoute": {
            "type": "string"
          },
          "ascentStyle": {
            "type": "string",
            "enum": [
              "onsight",
              "flash",
              "retro_flash",
              "redpoint",
              "toprope",
              "attempt"
            ]
          },
          "climbedAt": {
            "type": "string",
            "description": "ISO date; defaults to today"
          },
          "attempts": {
            "type": "number",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["idRoute", "ascentStyle"]
      }
    }
  }
}